:recycle: (EighteenTimeline) Extract pastStep
