+++ src/features/tutorial/helpers/useEighteenFirstStepperProps.tsx
@@ -2,18 +2,13 @@ import React from 'react'
 import styled from 'styled-components/native'
 
 import { CreditComponentProps } from 'features/tutorial/components/CreditTimeline'
+import {
+  PastStepBlock,
+  pastStep,
+} from 'features/tutorial/components/profileTutorial/Timelines/PastSteps'
 import { useDepositAmountsByAge } from 'shared/user/useDepositAmountsByAge'
-import { BicolorUnlock } from 'ui/svg/icons/BicolorUnlock'
 import { Warning } from 'ui/svg/icons/BicolorWarning'
-import { Spacer, Typo, getSpacing } from 'ui/theme'
-
-type Underage = 15 | 16 | 17
-
-const pastStep = (age: Underage, deposit: string): CreditComponentProps => ({
-  creditStep: 'pastStep',
-  children: <PastStepBlock text={`Tu as reçu ${deposit} à ${age}\u00a0ans`} />,
-  iconComponent: <GreyUnlock />,
-})
+import { getSpacing } from 'ui/theme'
 
 export const useEighteenFirstStepperProps = (
   activatedAt: number | null | undefined
@@ -63,25 +58,3 @@ const GreyWarning = styled(Warning).attrs(({ theme }) => ({
 }))({
   marginHorizontal: getSpacing(1.5),
 })
-
-const GreyUnlock = styled(BicolorUnlock).attrs(({ theme }) => ({
-  color: theme.colors.greySemiDark,
-  color2: theme.colors.greySemiDark,
-  size: theme.icons.sizes.smaller,
-}))({
-  marginHorizontal: getSpacing(1.5),
-})
-
-const StyledCaption = styled(Typo.Caption)(({ theme }) => ({
-  color: theme.colors.greyDark,
-}))
-
-const PastStepBlock = ({ text }: { text: string }) => {
-  return (
-    <React.Fragment>
-      <Spacer.Column numberOfSpaces={3} />
-      <StyledCaption>{text}</StyledCaption>
-      <Spacer.Column numberOfSpaces={1} />
-    </React.Fragment>
-  )
-}
